+++ src/main/java/rocks/cleancode/hamcrest/throwable/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Hamcrest matchers for {@link java.lang.Throwable}.
+ */
+package rocks.cleancode.hamcrest.throwable;