+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 
 ### Features
 
+- Matcher for `java.lang.Throwable`
+
 [Unreleased]: https://github.com/clean-code-rocks/hamcrest-java-throwable/commits/main